Verificar se arquivo existe antes de excluir

+++ src/controllers/UsuarioController.js
@@ -1,7 +1,7 @@
 const Usuario = require("../models/Usuario");
 const multer = require("multer");
 const path = require("path");
-const { unlink } = require("fs");
+const { existsSync, unlink } = require("fs");
 
 const show = async (req, res) => {
   try {
@@ -62,7 +62,7 @@ const upload = (req, res) => {
       // delete old image if exists
       const oldData = await Usuario.findById(req.usuario._id).select("imagem");
 
-      if (oldData.imagem) {
+      if (oldData.imagem && existsSync(`public/${oldData.imagem}`)) {
         unlink(`public/${oldData.imagem}`, (error) => {
           if (error) throw error;
         });
